Revert "Fixed double->single."
This reverts commit ec39a809bf82bae88eb6af292e9f0413055f70b1.

+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/ImageCropper/AspectRatioConfig.cs
@@ -8,6 +8,6 @@ namespace Microsoft.Toolkit.Uwp.SampleApp.SamplePages
     {
         public string Name { get; set; }
 
-        public float? AspectRatio { get; set; }
+        public double? AspectRatio { get; set; }
     }
 }
